Improve UBSignal initialization safety
UBSignal can be initialized with `init` however this is not always what someone wants to do, so to make this a little more explicit and safe, I'm marking `init` as unavailable, raising a runtime exception, and improving the headerdoc to denote that if you want an empty signal then you should pass in the `EmptySignal` protocol explicitly.

Also bumping the major to 2.0 as this is a breaking change.